requirements.txt
python-telegram-bot==20.7
aiohttp==3.8.5
python-dotenv==1.0.0
sqlalchemy==2.0.25